refactor(py): restructure into packages (models, storage, cli)
